Implement InitializeComponent
Changed the implementation to provide a value for `applicableToSpan` in
the `CompletionStartData` constructor. Theorized this would cause
actions like completion by double click to replace the text which
initially triggered the completion rather than having the text
pre-pended.

This was not the case. The value provided does not seem to stick even to
the immediate following call to `GetCompletionContextAsync`. By then the
span has already been cleared (the length is zero).